refactor: in progress organization
