+++ ethers-providers/src/rpc/provider.rs
@@ -1,18 +1,20 @@
+use ethers_core::types::SyncingStatus;
+
 use crate::{
     call_raw::CallBuilder,
-    ens, erc,
     errors::ProviderError,
-    maybe,
+    ext::{ens, erc},
     pubsub::{PubsubClient, SubscriptionStream},
     stream::{FilterWatcher, DEFAULT_LOCAL_POLL_INTERVAL, DEFAULT_POLL_INTERVAL},
+    utils::maybe,
     Http as HttpProvider, JsonRpcClient, JsonRpcClientWrapper, LogQuery, MiddlewareError,
-    MockProvider, NodeInfo, PeerInfo, PendingTransaction, QuorumProvider, RwClient, SyncingStatus,
+    MockProvider, NodeInfo, PeerInfo, PendingTransaction, QuorumProvider, RwClient,
 };
 
 #[cfg(all(not(target_arch = "wasm32"), feature = "ws"))]
-use crate::transports::Authorization;
+use crate::Authorization;
 #[cfg(not(target_arch = "wasm32"))]
-use crate::transports::{HttpRateLimitRetryPolicy, RetryClient};
+use crate::{HttpRateLimitRetryPolicy, RetryClient};
 
 #[cfg(feature = "celo")]
 use crate::CeloMiddleware;
@@ -1667,191 +1669,6 @@ impl ProviderExt for Provider<HttpProvider> {
 #[inline]
 pub fn is_local_endpoint(url: &str) -> bool {
     url.contains("127.0.0.1") || url.contains("localhost")
-}
-
-/// A middleware supporting development-specific JSON RPC methods
-///
-/// # Example
-///
-///```
-/// use ethers_providers::{Provider, Http, Middleware, DevRpcMiddleware};
-/// use ethers_core::types::TransactionRequest;
-/// use ethers_core::utils::Anvil;
-/// use std::convert::TryFrom;
-///
-/// # #[tokio::main(flavor = "current_thread")]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// let anvil = Anvil::new().spawn();
-/// let provider = Provider::<Http>::try_from(anvil.endpoint()).unwrap();
-/// let client = DevRpcMiddleware::new(provider);
-///
-/// // snapshot the initial state
-/// let block0 = client.get_block_number().await.unwrap();
-/// let snap_id = client.snapshot().await.unwrap();
-///
-/// // send a transaction
-/// let accounts = client.get_accounts().await?;
-/// let from = accounts[0];
-/// let to = accounts[1];
-/// let balance_before = client.get_balance(to, None).await?;
-/// let tx = TransactionRequest::new().to(to).value(1000).from(from);
-/// client.send_transaction(tx, None).await?.await?;
-/// let balance_after = client.get_balance(to, None).await?;
-/// assert_eq!(balance_after, balance_before + 1000);
-///
-/// // revert to snapshot
-/// client.revert_to_snapshot(snap_id).await.unwrap();
-/// let balance_after_revert = client.get_balance(to, None).await?;
-/// assert_eq!(balance_after_revert, balance_before);
-/// # Ok(())
-/// # }
-/// ```
-#[cfg(feature = "dev-rpc")]
-pub mod dev_rpc {
-    use crate::{Middleware, MiddlewareError, ProviderError};
-    use async_trait::async_trait;
-    use ethers_core::types::U256;
-    use thiserror::Error;
-
-    use std::fmt::Debug;
-
-    #[derive(Clone, Debug)]
-    pub struct DevRpcMiddleware<M>(M);
-
-    #[derive(Error, Debug)]
-    pub enum DevRpcMiddlewareError<M: Middleware> {
-        #[error("{0}")]
-        MiddlewareError(M::Error),
-
-        #[error("{0}")]
-        ProviderError(ProviderError),
-
-        #[error("Could not revert to snapshot")]
-        NoSnapshot,
-    }
-
-    #[async_trait]
-    impl<M: Middleware> Middleware for DevRpcMiddleware<M> {
-        type Error = DevRpcMiddlewareError<M>;
-        type Provider = M::Provider;
-        type Inner = M;
-
-        fn inner(&self) -> &M {
-            &self.0
-        }
-    }
-
-    impl<M: Middleware> MiddlewareError for DevRpcMiddlewareError<M> {
-        type Inner = M::Error;
-
-        fn from_err(src: M::Error) -> DevRpcMiddlewareError<M> {
-            DevRpcMiddlewareError::MiddlewareError(src)
-        }
-
-        fn as_inner(&self) -> Option<&Self::Inner> {
-            match self {
-                DevRpcMiddlewareError::MiddlewareError(e) => Some(e),
-                _ => None,
-            }
-        }
-    }
-
-    impl<M> From<ProviderError> for DevRpcMiddlewareError<M>
-    where
-        M: Middleware,
-    {
-        fn from(src: ProviderError) -> Self {
-            Self::ProviderError(src)
-        }
-    }
-
-    impl<M: Middleware> DevRpcMiddleware<M> {
-        pub fn new(inner: M) -> Self {
-            Self(inner)
-        }
-
-        // Ganache, Hardhat and Anvil increment snapshot ID even if no state has changed
-        pub async fn snapshot(&self) -> Result<U256, DevRpcMiddlewareError<M>> {
-            self.provider().request::<(), U256>("evm_snapshot", ()).await.map_err(From::from)
-        }
-
-        pub async fn revert_to_snapshot(&self, id: U256) -> Result<(), DevRpcMiddlewareError<M>> {
-            let ok = self
-                .provider()
-                .request::<[U256; 1], bool>("evm_revert", [id])
-                .await
-                .map_err(DevRpcMiddlewareError::ProviderError)?;
-            if ok {
-                Ok(())
-            } else {
-                Err(DevRpcMiddlewareError::NoSnapshot)
-            }
-        }
-    }
-    #[cfg(test)]
-    // Celo blocks can not get parsed when used with Ganache
-    #[cfg(not(feature = "celo"))]
-    mod tests {
-        use super::*;
-        use crate::{Http, Provider};
-        use ethers_core::utils::Anvil;
-        use std::convert::TryFrom;
-
-        #[tokio::test]
-        async fn test_snapshot() {
-            let anvil = Anvil::new().spawn();
-            let provider = Provider::<Http>::try_from(anvil.endpoint()).unwrap();
-            let client = DevRpcMiddleware::new(provider);
-
-            // snapshot initial state
-            let block0 = client.get_block_number().await.unwrap();
-            let time0 = client.get_block(block0).await.unwrap().unwrap().timestamp;
-            let snap_id0 = client.snapshot().await.unwrap();
-
-            // mine a new block
-            client.provider().mine(1).await.unwrap();
-
-            // snapshot state
-            let block1 = client.get_block_number().await.unwrap();
-            let time1 = client.get_block(block1).await.unwrap().unwrap().timestamp;
-            let snap_id1 = client.snapshot().await.unwrap();
-
-            // mine some blocks
-            client.provider().mine(5).await.unwrap();
-
-            // snapshot state
-            let block2 = client.get_block_number().await.unwrap();
-            let time2 = client.get_block(block2).await.unwrap().unwrap().timestamp;
-            let snap_id2 = client.snapshot().await.unwrap();
-
-            // mine some blocks
-            client.provider().mine(5).await.unwrap();
-
-            // revert_to_snapshot should reset state to snap id
-            client.revert_to_snapshot(snap_id2).await.unwrap();
-            let block = client.get_block_number().await.unwrap();
-            let time = client.get_block(block).await.unwrap().unwrap().timestamp;
-            assert_eq!(block, block2);
-            assert_eq!(time, time2);
-
-            client.revert_to_snapshot(snap_id1).await.unwrap();
-            let block = client.get_block_number().await.unwrap();
-            let time = client.get_block(block).await.unwrap().unwrap().timestamp;
-            assert_eq!(block, block1);
-            assert_eq!(time, time1);
-
-            // revert_to_snapshot should throw given non-existent or
-            // previously used snapshot
-            let result = client.revert_to_snapshot(snap_id1).await;
-            assert!(result.is_err());
-
-            client.revert_to_snapshot(snap_id0).await.unwrap();
-            let block = client.get_block_number().await.unwrap();
-            let time = client.get_block(block).await.unwrap().unwrap().timestamp;
-            assert_eq!(block, block0);
-            assert_eq!(time, time0);
-        }
-    }
 }
 
 #[cfg(test)]
